requirements.txt
# API dependencies
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
sqlalchemy==2.0.12
psycopg2-binary==2.9.6
alembic==1.10.4
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
redis==4.5.5

# Scanner dependencies
docker==6.0.1
boto3==1.26.135
requests==2.30.0
pyyaml==6.0
trivy-python==0.1.3

# Policy engine dependencies
opa-python==1.2.0
jsonschema==4.17.3

# Testing dependencies
pytest==7.3.1
pytest-cov==4.1.0
httpx==0.24.0
pytest-mock==3.10.0

# Security dependencies
cryptography==40.0.2
EOF < /dev/null